requirements.txt
# Web framework
fastapi==0.115.0
uvicorn==0.30.0
starlette==0.41.0

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Validation
pydantic==2.10.0
pydantic-settings==2.6.0
email-validator==2.2.0

# HTTP
httpx==0.28.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
aiosqlite==0.20.0

# Utils
python-dotenv==1.0.1
